# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Yanki Wallet API
  version: 1.0.0
servers:
  - url: http://localhost:8091
paths:
  /v1/wallets:
    get:
      summary: Retrieve all wallets
      operationId: getAllWallets
      responses:
        '200':
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WalletResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a wallet
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRequest'
      responses:
        '201':
          description: Card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/wallets/{walletId}:
    get:
      summary: Retrieve a wallet by ID
      operationId: getWalletById
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update
      operationId: updateWallet
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRequest'
      responses:
        '200':
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          '404':
            description: Wallet not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a wallet
      operationId: deleteWallet
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Wallet deleted
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/wallets/{phone}/balance:
    get:
      summary: Get wallet balances
      operationId: getPrimaryWalletBalance
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          description: Phone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/wallets/{walletId}/associate:
    post:
      summary: Associate a wallet to a debit card
      operationId: associateDebitCard
      tags:
        - Wallets
      parameters:
        - name: walletId
          in: path
          required: true
          description: Wallet ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociateRequest'
      responses:
        '200':
          description: Wallet successfully associated with the debit card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
          '400':
            description: Malformed Wallet data
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '404':
            description: Wallet not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
  /v1/wallets/yankear:
    post:
      summary: Perform a money transfer between two users via yankear
      description: Allows a user to send money to another user via their phone number
      operationId: yankearWallet
      tags:
        - Wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YankearRequest'
      responses:
        '200':
          description: Yankear transfer successfully completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YankearResponse'
          '400':
            description: Bad request error, such as insufficient balance or invalid data
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '404':
            description: User or wallet not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
  /v1/wallets/{walletId}/transactions:
    get:
      summary: Get wallet transactions
      operationId: getWalletTransactions
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wallet transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    WalletRequest:
      type: object
      required:
        - name
        - dni
        - phone
      properties:
        name:
          type: string
        documentNumber:
          type: string
        documentType:
          type: string
        phone:
          type: string
        imei:
          type: string
        email:
          type: string
    WalletResponse:
      type: object
      required:
        - name
        - dni
        - phone
        - debitCardId
      properties:
        id:
          type: string
        name:
          type: string
        documentNumber:
          type: string
        documentType:
          type: string
        phone:
          type: string
        imei:
          type: string
        email:
          type: string
        debitCardId:
          type: string,
        balance:
          type: decimal
        transactions:
          type: array
    ErrorResponse:
      type: object
      required:
        - status
        - code
        - message
        - timestamp
      properties:
        status:
          type: int
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    BalanceResponse:
      type: object
      properties:
        walletId:
          type: string
        name:
          type: string
        phone:
          type: string
        balance:
          type: decimal
    AssociateRequest:
      type: object
      required:
        - debitCardId
      properties:
        debitCardId:
          type: string
    YankearRequest:
      type: object
      required:
        - fromPhoneNumber
        - toPhoneNumber
        - amount
      properties:
        fromPhoneNumber:
          type: string
        toPhoneNumber:
          type: string
        amount:
          type: decimal
        description:
          type: string
    YankearResponse:
      type: object
      required:
        - name
        - PhoneNumber
        - amount
        - description
        - timestamp
      properties:
        name:
          type: string
        PhoneNumber:
          type: string
        amount:
          type: decimal
        description:
          type: string
        timestamp:
          type: string
          format : date-time